+++ src/App.tsx
@@ -5,8 +5,9 @@ import Bar from './components/Bar';
 import Steps from './components/Stepper/Steps';
 import { UserCtx } from './contexts/user.context';
 import { useLocalStorage } from './hooks/localstorage.hook';
+import { User } from './models';
 import theme from './theme/theme';
-import { createUserFromLocalStorage, User } from './types';
+import { createUserFromLocalStorage } from './transformers/user.transformer';
 
 const hasTokenExpired = (userObj: User): boolean => {
   return userObj && userObj.token && new Date() > userObj.token.expiresAt;
