+++ src/components/Auth/Login.tsx
@@ -1,7 +1,8 @@
 import React, { memo, useContext } from 'react';
 import GoogleLogin from 'react-google-login';
 import { UserCtx } from '../../contexts/user.context';
-import { createUserFromJson, User } from '../../types';
+import { User } from '../../models';
+import { createUserFromJson } from '../../transformers/user.transformer';
 
 const CLIENT_ID = process.env.REACT_APP_GOOGLE_ID || '';
