+++ src/components/SendEmailButton.tsx
@@ -1,8 +1,8 @@
 import Button from '@material-ui/core/Button';
 import React, { FunctionComponent, memo, useContext } from 'react';
 import { UserCtx } from '../contexts/user.context';
+import { Recipient } from '../models';
 import send from '../services/MailSender';
-import { Recipient } from '../types';
 
 interface OwnProps {
   rcps: Recipient[];
