+++ src/models/EmailMeta.ts
