+++ src/models/User.ts
@@ -1,0 +1,10 @@
+import {UserToken} from "./UserToken";
+
+export interface User {
+  email: string;
+  firstName: string;
+  lastName: string;
+  googleId: number;
+  name: string;
+  token: UserToken;
+}
