+++ src/models/UserToken.ts
@@ -1,0 +1,5 @@
+export interface UserToken {
+  accessToken: string;
+  expiresAt: Date;
+  idToken: string;
+}
