+++ src/transformers/spreadsheet.transformer.ts
@@ -1,4 +1,4 @@
-import { Spreadsheet } from '../types';
+import { Spreadsheet } from '../models';
 
 export class SpreadSheetTransformer {
   public transform(data: any): Spreadsheet {
