--- src/types/index.ts
@@ -1,6 +1,0 @@
-export * from './email';
-export * from './nullable';
-export * from './recipient';
-export * from './spreadsheet';
-export * from './step';
-export * from './user';
